* Adaptation to new API in MPXDJ plugin * Rendering changes in all 3 main views * Compilation fix in library-mlibman.cc

+++ src/library.hh
@@ -190,7 +190,7 @@ namespace MPX
             markovUpdate(gint64 /* track a */, gint64 /* track b */) ;
 
             gint64 
-            markovGetRandomProbableTrack(gint64 /* track a*/); 
+            markovGetRandomProbableTrack(int /* track a*/); 
 
             gint64
             collectionCreate(const std::string& /*name*/, const std::string& /*blurb*/) ;
